+++ lib/marskal/core_ext.rb
@@ -20,6 +20,7 @@ require "marskal/core_ext/date"
 require "marskal/core_ext/file"
 require "marskal/core_ext/hash"
 require "marskal/core_ext/i18n"
+require "marskal/core_ext/string"
 # require "marskal/core_ext/my_mod"  #used for testing, take out eventually