+++ test/marskal/marskal_core_ext_my_date_test.rb
@@ -7,7 +7,7 @@ require 'test_helper'
 # * <tt>Created: 2016-10-2016</tt> <b>Mike Urban</b> <mike@marskalgroup.com>
 #
 # ---
-describe 'Tests for Marskal::CoreExt::MyDate' do
+describe 'Marskal::CoreExt::MyDate' do
 
   ##
   # Prepare variables for testing